added tests to message and conversation functionnality
