+++ ventashop/message_views.py
@@ -20,11 +20,11 @@ class MessageListView(ListView, FormMixin):
     def get_queryset(self, *args, **kwargs):
         """Conversation message list to be displayed."""
 
-        if "last" in self.kwargs:
-            print("IN IF CONDITION")
+        if "last" in self.kwargs:   # n last messages
             last = int(self.kwargs["last"])
             m_set = list(Message.objects.filter(conversation__id=self.kwargs["pk"]))[-last:]
-        else:
+        
+        else:                       # all messages
             m_set = Message.objects.filter(conversation__id=self.kwargs["pk"])
 
         return m_set
@@ -36,12 +36,12 @@ class MessageListView(ListView, FormMixin):
         conversation = get_object_or_404(Conversation, pk=self.kwargs["pk"])
         
         if form.is_valid():
-            author = "Tom"
+            author = "Tom"      # Hardcoded for now, user mechanism is not yet implemented.
             content = form.cleaned_data["content"]
 
             conversation.add_message(author=author, content=content)
 
-        return HttpResponseRedirect(reverse('ventashop:messages', args=(conversation.id,)))
+        return HttpResponseRedirect(reverse('ventashop:messages-last', args=(conversation.id, 5)))
         
     def get_context_data(self, **kwargs):
         """Related conversation and new message form to be displayed."""
@@ -56,6 +56,6 @@ class ConversationListView(ListView):
     """Our conversation list view."""
 
     model = Conversation
-    paginate_by = 5  # if pagination is desired
+    # paginate_by = 5  # if pagination is desired
     template_name = 'ventashop/conversations.html'
     context_object_name = "conversation_list"
